+++ pygame/controllers/__init__.py
@@ -1,0 +1,5 @@
+
+"""Controllers package"""
+
+from . import GameLoopState
+from . import ControllersManager
